+++ Source/ParseResult.swift
@@ -16,10 +16,10 @@
  */
 
 public enum ParseResult<Output> {
-
-    case success(result: Output, tail: TokenStream)
+    // swiftlint:disable identifier_name // Triggers a false positive in swift lint.
+    case success(result: Output, tail: Substring)
     case failure(details: ParseFailure)
-
+    // swiftlint:enable identifier_name
 }
 
 /**
@@ -55,9 +55,10 @@ public extension ParseResult {
 public enum ParseFailure: Equatable {
 
     case insufficiantTokens
-    case unexpectedToken(token: TokenStream, tail: TokenStream)
+    // swiftlint:disable identifier_name // Triggers a false positive in swift lint.
+    case unexpectedToken(token: Character, tail: Substring)
     case custom(message: String)
-
+    // swiftlint:enable identifier_name
 }
 
 /**
@@ -73,7 +74,7 @@ public func == (left: ParseFailure, right: ParseFailure) -> Bool {
     case (.insufficiantTokens, .insufficiantTokens):
         return true
     case (.unexpectedToken(let leftToken, let leftTail), .unexpectedToken(let rightToken, let rightTail)):
-        return String(leftToken) == String(rightToken) && String(leftTail) == String(rightTail)
+        return leftToken == rightToken && leftTail == rightTail
     case (.custom(let leftMesage), .custom(let rightMessage)):
         return leftMesage == rightMessage
     default: