+++ Tests/ParserCombinatorTests/CharacterParserTests+PremadeParsers.swift
@@ -19,42 +19,42 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: "A") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "b") {
             XCTAssertEqual("b", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "AB") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("B", String(results.tail))
+            XCTAssertEqual("B", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "A1234") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("1234", String(results.tail))
+            XCTAssertEqual("1234", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "1234") {
-            XCTAssertEqual("1", String(unexpected.token))
-            XCTAssertEqual("234", String(unexpected.tail))
+            XCTAssertEqual("1", unexpected.token)
+            XCTAssertEqual("234", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "1") {
-            XCTAssertEqual("1", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("1", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
@@ -68,35 +68,35 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: "a") {
             XCTAssertEqual("a", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "ab") {
             XCTAssertEqual("a", results.result)
-            XCTAssertEqual("b", String(results.tail))
+            XCTAssertEqual("b", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "aBCD") {
             XCTAssertEqual("a", results.result)
-            XCTAssertEqual("BCD", String(results.tail))
+            XCTAssertEqual("BCD", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "Abcd") {
-            XCTAssertEqual("A", String(unexpected.token))
-            XCTAssertEqual("bcd", String(unexpected.tail))
+            XCTAssertEqual("A", unexpected.token)
+            XCTAssertEqual("bcd", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "A") {
-            XCTAssertEqual("A", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("A", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
@@ -110,35 +110,35 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: "A") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "AB") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("B", String(results.tail))
+            XCTAssertEqual("B", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "Abcd") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("bcd", String(results.tail))
+            XCTAssertEqual("bcd", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "aBCD") {
-            XCTAssertEqual("a", String(unexpected.token))
-            XCTAssertEqual("BCD", String(unexpected.tail))
+            XCTAssertEqual("a", unexpected.token)
+            XCTAssertEqual("BCD", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "a") {
-            XCTAssertEqual("a", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("a", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
@@ -152,42 +152,42 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: "A") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "1") {
             XCTAssertEqual("1", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "A1") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("1", String(results.tail))
+            XCTAssertEqual("1", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "A#$%") {
             XCTAssertEqual("A", results.result)
-            XCTAssertEqual("#$%", String(results.tail))
+            XCTAssertEqual("#$%", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "#BCD") {
-            XCTAssertEqual("#", String(unexpected.token))
-            XCTAssertEqual("BCD", String(unexpected.tail))
+            XCTAssertEqual("#", unexpected.token)
+            XCTAssertEqual("BCD", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "#") {
-            XCTAssertEqual("#", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("#", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
@@ -201,28 +201,28 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: "1") {
             XCTAssertEqual("1", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "1#$%") {
             XCTAssertEqual("1", results.result)
-            XCTAssertEqual("#$%", String(results.tail))
+            XCTAssertEqual("#$%", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "#123") {
-            XCTAssertEqual("#", String(unexpected.token))
-            XCTAssertEqual("123", String(unexpected.tail))
+            XCTAssertEqual("#", unexpected.token)
+            XCTAssertEqual("123", unexpected.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "A") {
-            XCTAssertEqual("A", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("A", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
@@ -237,28 +237,28 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: " ") {
             XCTAssertEqual(" ", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: " #$%") {
             XCTAssertEqual(" ", results.result)
-            XCTAssertEqual("#$%", String(results.tail))
+            XCTAssertEqual("#$%", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "#   ") {
-            XCTAssertEqual("#", String(unexpected.token))
-            XCTAssertEqual("   ", String(unexpected.tail))
+            XCTAssertEqual("#", unexpected.token)
+            XCTAssertEqual("   ", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "A") {
-            XCTAssertEqual("A", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("A", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
@@ -271,28 +271,28 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: "\n") {
             XCTAssertEqual("\n", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "\n#$%") {
             XCTAssertEqual("\n", results.result)
-            XCTAssertEqual("#$%", String(results.tail))
+            XCTAssertEqual("#$%", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "#\n") {
-            XCTAssertEqual("#", String(unexpected.token))
-            XCTAssertEqual("\n", String(unexpected.tail))
+            XCTAssertEqual("#", unexpected.token)
+            XCTAssertEqual("\n", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "A") {
-            XCTAssertEqual("A", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("A", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
@@ -306,35 +306,35 @@ extension CharacterParserTests {
 
         if case .success(let results) = parserUnderTest.run(withInput: " ") {
             XCTAssertEqual(" ", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "\n") {
             XCTAssertEqual("\n", results.result)
-            XCTAssertEqual("", String(results.tail))
+            XCTAssertEqual("", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if case .success(let results) = parserUnderTest.run(withInput: "\n#$%") {
             XCTAssertEqual("\n", results.result)
-            XCTAssertEqual("#$%", String(results.tail))
+            XCTAssertEqual("#$%", results.tail)
         } else {
             XCTFail("The parser should succeed")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "#\n ") {
-            XCTAssertEqual("#", String(unexpected.token))
-            XCTAssertEqual("\n ", String(unexpected.tail))
+            XCTAssertEqual("#", unexpected.token)
+            XCTAssertEqual("\n ", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }
 
         if let unexpected = ParserTestHelper.findUnexpectedToken(running: parserUnderTest, with: "A") {
-            XCTAssertEqual("A", String(unexpected.token))
-            XCTAssertEqual("", String(unexpected.tail))
+            XCTAssertEqual("A", unexpected.token)
+            XCTAssertEqual("", unexpected.tail)
         } else {
             XCTFail("The parser should find an unexpected token")
         }