+++ CHANGELOG.md
@@ -38,3 +38,11 @@
 * Swift 4 and Xcode 9 updates.
 * Swiftlint version update to 0.22.0
 * Add failure messages to all XCFail test cases.
+
+## Version 2.0.0
+
+### *Released 2019-09-01*
+
+* Swift 5 and Xcode 10.3 updates.
+* Swiftlint version update to 0.34.0
+* Add failure messages to all XCFail test cases.